Capture Earnings Management from Financial Restatements
In this study, we investigated the use of financial analysis techniques for fraud detection. We first used the Jones model and the Modified Jones model to determine if firms that issue financial restatements tend to have higher discretionary accruals than firms that do not issue financial restatements. Moreover, we explored other attributes that may predict fraud and used a logistic regression model to predict fraudulent firms based on their financial data. Finally, we verify our fraud prediction with 2018 restatement data.